Make PersonListViewCell an inner class
